Create movie-card.jsx
The MovieCard component will receive information about movies as its props and won't have a state. It will be a low-level, pure component. This is because you want to be able to reuse it later in other views. The less rendering it performs, the more generic it will be, and the more roles it will be able to fulfill.