adding .cache to to .gitignore
From the main myFlix folder (not from the “client” folder), push your changes to GitHub. Note: Remember that your GitHub repository is tied to your main myFlix folder—not with the “client” folder. This is why you need to push the changes from the main folder and not from the “client” folder. The same reasoning goes to adding the “.cache” to to your “.gitignore” file in the main folder rather than creating a new one inside the “client” folder.
Submit a link to your GitHub repository